ci: add set as File for build in ci
