+++ src/models/actions/member.ts
@@ -1,21 +1,7 @@
 import { z } from "zod";
 
+import { FileType } from "@/lib/file";
 import { memberSchema } from "@/models/member";
-
-// Define a mock file class for environments without File API
-class MockFile {
-    size: number;
-    name: string;
-    type: string;
-    constructor(buffer: Buffer, name: string, type: string) {
-        this.size = buffer.length;
-        this.name = name;
-        this.type = type;
-    }
-}
-
-// Use the global File if available, otherwise use MockFile
-const FileType = typeof File !== "undefined" ? File : MockFile;
 
 export const memberInfoUpdateSchema = z.object({
     member: z.object({
